@@ -18,13 +18,13 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   get_storage: ^2.0.3
-  http: ^0.12.0+2
+  http: ^0.13.3
   unorm_dart: ^0.2.0
 
   # The following adds the Cupertino Icons font to your application.
